backend/requirements.txt
Flask==2.1.0
web3==5.31.0
python-dotenv==0.20.0
